refactor: extract spaces feature folder
Move Space schema, lib, dialogs, and space-list + space-settings
screens into src/app/features/spaces/. Schema and createSpace
re-exported from @/schema for compatibility.
