+++ src/app/features/spaces/lib/spaces.ts
@@ -1,7 +1,8 @@
 import { Group, co, type ID } from "jazz-tools"
-import { Space, UserAccount } from "@/schema"
+import { UserAccount } from "@/schema"
+import { Space } from "./schema"
 import { permanentlyDeleteSpace as deleteSpaceCoValue } from "@/lib/delete-covalue"
-import { buildSpaceInviteLink } from "@/app/features/sharing"
+import { buildSpaceInviteLink } from "@/app/features/sharing/lib/invite-links"
 
 export {
 	createSpaceInvite,
