+++ src/app/features/spaces/widgets/space-selector.tsx
@@ -35,7 +35,7 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { UserAccount, Space, createSpace } from "@/schema"
-import { getSpaceGroup } from "@/lib/spaces"
+import { getSpaceGroup } from "../lib/spaces"
 import { testIds } from "@/lib/test-ids"
 
 export { SpaceSelector, SpaceInitials }
