+++ src/schema/document.ts
@@ -1,0 +1,54 @@
+import { co, z } from "jazz-tools"
+
+export {
+	ImageAsset,
+	VideoAsset,
+	Asset,
+	Document,
+	CursorEntry,
+	CursorFeed,
+	HighlightRange,
+}
+
+let CursorEntry = z.object({
+	position: z.number(),
+	selectionEnd: z.number().optional(),
+})
+
+let CursorFeed = co.feed(CursorEntry)
+
+let ImageAsset = co.map({
+	type: z.literal("image"),
+	name: z.string(),
+	image: co.image(),
+	createdAt: z.date(),
+})
+
+let VideoAsset = co.map({
+	type: z.literal("video"),
+	name: z.string(),
+	video: co.fileStream(),
+	mimeType: z.string(),
+	muteAudio: z.boolean().optional(),
+	createdAt: z.date(),
+})
+
+let Asset = co.discriminatedUnion("type", [ImageAsset, VideoAsset])
+
+let HighlightRange = z.object({
+	start: z.number(),
+	end: z.number(),
+})
+
+let Document = co.map({
+	version: z.literal(1),
+	content: co.plainText(),
+	assets: co.optional(co.list(Asset)),
+	cursors: co.optional(CursorFeed),
+	deletedAt: z.date().optional(),
+	presentationLine: z.number().optional(),
+	highlightRange: HighlightRange.optional(),
+	spaceId: z.string().optional(),
+	createdAt: z.date(),
+	updatedAt: z.date(),
+})
